+++ tests/unit/controller/CbfController_test.py
@@ -67,12 +67,20 @@ class TestCbfController:
                 "mid_csp_cbf/vcc/002",
                 "mid_csp_cbf/vcc/003",
                 "mid_csp_cbf/vcc/004",
+                "mid_csp_cbf/vcc/005",
+                "mid_csp_cbf/vcc/006",
+                "mid_csp_cbf/vcc/007",
+                "mid_csp_cbf/vcc/008",
             ],
             FSP=[
                 "mid_csp_cbf/fsp/01",
                 "mid_csp_cbf/fsp/02",
                 "mid_csp_cbf/fsp/03",
                 "mid_csp_cbf/fsp/04",
+                "mid_csp_cbf/fsp/05",
+                "mid_csp_cbf/fsp/06",
+                "mid_csp_cbf/fsp/07",
+                "mid_csp_cbf/fsp/08",
             ],
             TalonLRU=[
                 "mid_csp_cbf/talon_lru/001",
@@ -101,11 +109,11 @@ class TestCbfController:
             FsSLIMConfigPath="mnt/slim/fs/slim_config.yaml",
             VisSLIMConfigPath="mnt/slim/vis/slim_config.yaml",
             LRCTimeout="30",
-            MaxCapabilities=["VCC:8", "FSP:4", "Subarray:1"],
+            MaxCapabilities=["VCC:8", "FSP:8", "Subarray:1"],
         )
 
         for name, mock in initial_mocks.items():
-            harness.add_mock_device(device_name=name, device_mock=mock)
+            harness.add_mock_device(device_name=name, device_mock=mock(name))
 
         with harness as test_context:
             yield test_context
@@ -297,7 +305,6 @@ class TestCbfController:
             ),
         )
 
-    @pytest.mark.skip(reason="Skipping test involving nested LRC")
     def test_Commands_all(
         self: TestCbfController,
         device_under_test: context.DeviceProxy,
@@ -311,7 +318,24 @@ class TestCbfController:
                              events from the device under test.
         """
         # Establish communication
-        self.test_Online(device_under_test, event_tracer)
+        # TODO: mock talon configuration to test with sim mode FALSE
+        # device_under_test.simulationMode = SimulationMode.FALSE
+        device_under_test.adminMode = AdminMode.ONLINE
+        assert_that(event_tracer).within_timeout(
+            test_utils.EVENT_TIMEOUT
+        ).has_change_event_occurred(
+            device_name=device_under_test,
+            attribute_name="adminMode",
+            attribute_value=AdminMode.ONLINE,
+        )
+
+        assert_that(event_tracer).within_timeout(
+            test_utils.EVENT_TIMEOUT
+        ).has_change_event_occurred(
+            device_name=device_under_test,
+            attribute_name="state",
+            attribute_value=DevState.OFF,
+        )
 
         with open(json_file_path + "sys_param_4_boards.json") as f:
             sp = f.read()
@@ -340,6 +364,8 @@ class TestCbfController:
                         f"{return_value[1][0]}",
                         f'[{ResultCode.OK.value}, "{command_name} completed OK"]',
                     ),
+                    None,
+                    1,
                 )
             )